require only --server-yaml to use server native yaml format, move server native yaml code to its own package
